[FIXED] generating indices of FBX files

+++ Core/src/RabBit/entity/components/Mesh.h
@@ -10,7 +10,7 @@ namespace RB::Entity
     public:
 
         Mesh(const char* file_name);
-        Mesh(const char* name, float* vertex_data, uint32_t elements_per_vertex, uint64_t vertex_data_count, uint32_t* index_data, uint64_t index_data_count);
+        Mesh(const char* name, float* vertex_data, uint32_t elements_per_vertex, uint64_t vertex_data_count, uint16_t* index_data, uint64_t index_data_count);
 
         ~Mesh()
         {
